requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
aiosqlite==0.20.0
google-generativeai==0.5.4
pillow==10.3.0
python-multipart==0.0.9
pydantic-settings==2.2.1
python-dotenv==1.0.1
aiofiles==23.2.1